requirements.txt
django-nested-admin
django-crispy-forms
Pillow
pandas
openpyxl
mysqlclient
openai
python-dotenv
